fix logging and some includes
