+++ src/py21cmfast/src/21cmFAST.h
@@ -1,56 +1,53 @@
 /*
     This is the header file for the wrappable version of 21cmFAST, or 21cmMC.
-    It contains all the header files of the backend.
+    It contains all the header files of the backend. This is not imported by
+    any module in the backend, but is provided to the cffi set_source() call,
+    and makes sure all headers are accessible to the compiler.
 */
 
-// #include <stdlib.h>
-// #include <stdio.h>
-// #include <math.h>
-// #include <complex.h>
-// #include <fftw3.h>
-// #include <gsl/gsl_integration.h>
-// #include <gsl/gsl_roots.h>
-// #include <gsl/gsl_interp.h>
-// #include <gsl/gsl_spline.h>
-// #include <gsl/gsl_errno.h>
-// #include <gsl/gsl_rng.h>
-// #include <gsl/gsl_randist.h>
+#include <stdlib.h>
+#include <stdio.h>
+#include <math.h>
+#include <complex.h>
+#include <fftw3.h>
+#include <gsl/gsl_integration.h>
+#include <gsl/gsl_roots.h>
+#include <gsl/gsl_interp.h>
+#include <gsl/gsl_spline.h>
+#include <gsl/gsl_errno.h>
+#include <gsl/gsl_rng.h>
+#include <gsl/gsl_randist.h>
 
 
 //  included in the sources keyword of ffi.set_source
-// #include "BrightnessTemperatureBox.h"
-// #include "bubble_helper_progs.h"
-// #include "cexcept.h"
-// #include "Constants.h"
-// #include "cosmology.h"
-// #include "debugging.h"
-// #include "dft.h"
-// #include "elec_interp.h"
-// #include "exceptions.h"
-// #include "filtering.h"
-// #include "HaloBox.h"
-// #include "HaloField.h"
-// #include "heating_helper_progs.h"
-// #include "hmf.h"
-// #include "indexing.h"
-// #include "InitialConditions.h"
-// #include "InputParameters.h"
-// #include "interp_tables.h"
-// #include "interpolation.h"
-// #include "IonisationBox.h"
-// #include "logger.h"
-// #include "LuminosityFunction.h"
-// #include "OutputStructs.h"
-// #include "PerturbField.h"
-// #include "PerturbHaloField.h"
-// #include "photoncons.h"
-// #include "recombinations.h"
-// #include "SpinTemperatureBox.h"
-// #include "Stochasticity.h"
-// #include "subcell_rsds.h"
-// #include "thermochem.h"
-
+#include "BrightnessTemperatureBox.h"
+#include "bubble_helper_progs.h"
+#include "cexcept.h"
+#include "Constants.h"
+#include "cosmology.h"
+#include "debugging.h"
+#include "dft.h"
+#include "elec_interp.h"
+#include "exceptions.h"
+#include "filtering.h"
+#include "HaloBox.h"
+#include "HaloField.h"
+#include "heating_helper_progs.h"
+#include "hmf.h"
+#include "indexing.h"
+#include "InitialConditions.h"
 #include "InputParameters.h"
-#include "OutputStructs.h"
-#include "photoncons.h"
+#include "interp_tables.h"
+#include "interpolation.h"
+#include "IonisationBox.h"
 #include "logger.h"
+#include "LuminosityFunction.h"
+#include "OutputStructs.h"
+#include "PerturbField.h"
+#include "PerturbHaloField.h"
+#include "photoncons.h"
+#include "recombinations.h"
+#include "SpinTemperatureBox.h"
+#include "Stochasticity.h"
+#include "subcell_rsds.h"
+#include "thermochem.h"
