+++ src/py21cmfast/src/hmf.c
@@ -12,10 +12,14 @@
 #include "Constants.h"
 #include "InputParameters.h"
 #include "cosmology.h"
+#include "interp_tables.h"
 
 #include "hmf.h"
 
 #define EPS2 3.0e-11 //small number limit for GL integration
+#define JENKINS_a (0.73) //Jenkins+01, SMT has 0.707
+#define JENKINS_b (0.34) //Jenkins+01 fit from Barkana+01, SMT has 0.5
+#define JENKINS_c (0.81) //Jenkins+01 from from Barkana+01, SMT has 0.6
 
 //Gauss-Legendre integration constants
 #define NGL_INT 100 // 100
@@ -32,6 +36,259 @@ double get_frac_limit(double M, double norm, double alpha, double limit, bool mi
 
     //if alpha is zero, this returns 1 as expected (note strict inequalities above)
     return pow(M/pivot,alpha);
+}
+
+////MASS FUNCTIONS BELOW//////
+
+/* sheth correction to delta crit */
+double sheth_delc_dexm(double del, double sig){
+    return sqrt(SHETH_a)*del*(1. + global_params.SHETH_b*pow(sig*sig/(SHETH_a*del*del), global_params.SHETH_c));
+}
+
+/*DexM uses a fit to this barrier to acheive MF similar to ST, Here I use the fixed version for the sampler*/
+//NOTE: if I made this a table it would save a pow call per condition in the sampler
+double sheth_delc_fixed(double del, double sig){
+    return sqrt(JENKINS_a)*del*(1. + JENKINS_b*pow(sig*sig/(JENKINS_a*del*del), JENKINS_c));
+}
+
+//Get the relevant excursion set barrier density given the user-specified HMF
+double get_delta_crit(int HMF, double sigma, double growthf){
+    if(HMF==4)
+        return DELTAC_DELOS;
+    if(HMF==1)
+        return sheth_delc_fixed(Deltac/growthf,sigma)*growthf;
+
+    return Deltac;
+}
+
+/*
+Unconditional Mass function from Delos 2023 (https://arxiv.org/pdf/2311.17986.pdf)
+Matches well with N-bodies (M200), has a corresponding Conditional Mass Function (below) and
+an excursion set method. Hence can be consistently used throughout the Halo Finder, Halo Sampler
+And radiation. The mass functions are based off a constant barrier delta = 1.5 and a top-hat window function
+*/
+double dNdlnM_Delos(double growthf, double lnM){
+    double dfdnu,dsigmadm,sigma,sigma_inv,dfdM;
+    double nu;
+    //hardcoded for now
+    const double coeff_nu = 0.519;
+    const double index_nu = 0.582;
+    const double exp_factor = -0.469;
+
+    sigma = EvaluateSigma(lnM);
+    sigma_inv = 1/sigma;
+    dsigmadm = EvaluatedSigmasqdm(lnM) * (0.5*sigma_inv); //d(s^2)/dm z0 to dsdm
+
+    nu = DELTAC_DELOS*sigma_inv/growthf;
+
+    dfdnu = coeff_nu*pow(nu,index_nu)*exp(exp_factor*nu*nu);
+    dfdM = dfdnu * fabs(dsigmadm) * sigma_inv;
+
+    //NOTE: dfdM == constants*dNdlnM
+    return dfdM*cosmo_params_global->OMm*RHOcrit;
+}
+
+double dNdlnM_conditional_Delos(double growthf, double lnM, double delta_cond, double sigma_cond){
+    double result,dfdnu,dsigmadm,sigma,sigdiff_inv,dfdM,sigma_inv;
+    double nu;
+    //hardcoded for now
+    const double coeff_nu = 0.519;
+    const double index_nu = 0.582;
+    const double exp_factor = -0.469;
+
+    sigma = EvaluateSigma(lnM);
+    if(sigma < sigma_cond) return 0.;
+    dsigmadm = EvaluatedSigmasqdm(lnM) * 0.5; //d(s^2)/dm to s*dsdm
+    sigdiff_inv = sigma == sigma_cond ? 1e6 : 1/(sigma*sigma - sigma_cond*sigma_cond);
+
+    nu = (DELTAC_DELOS - delta_cond)*sqrt(sigdiff_inv)/growthf;
+
+    dfdnu = coeff_nu*pow(nu,index_nu)*exp(exp_factor*nu*nu);
+    dfdM = dfdnu * fabs(dsigmadm) * sigdiff_inv;
+
+    //NOTE: like the other CMFs this is dNdlogM and leaves out
+    //   the (cosmo_params_global->OMm)*RHOcrit
+    //NOTE: dfdM == constants*dNdlnM
+    return dfdM;
+}
+
+//Sheth Tormen 2002 fit for the CMF, while the moving barrier does not allow for a simple rescaling, it has been found
+//That a taylor expansion of the barrier shape around the point of interest well approximates the simulations
+double st_taylor_factor(double sig, double sig_cond, double growthf, double *zeroth_order){
+    int i;
+    double a = JENKINS_a;
+    double alpha = JENKINS_c; //fixed instead of global_params.SHETH_c bc of DexM corrections
+    double beta = JENKINS_b; //fixed instead of global_params.SHETH_b
+
+    double del = Deltac/growthf;
+
+    double sigsq = sig*sig;
+    double sigsq_inv = 1./sigsq;
+    double sigcsq = sig_cond*sig_cond;
+    double sigdiff = sig == sig_cond ? 1e-6 : sigsq - sigcsq;
+
+    // This array cumulatively builds the taylor series terms
+    // sigdiff^n / n! * df/dsigma (polynomial w alpha)
+    double t_array[6];
+    t_array[0] = 1.;
+    for(i=1;i<6;i++)
+        t_array[i] = t_array[i-1] * (-sigdiff) / i * (alpha-i+1) * sigsq_inv;
+
+    //Sum small to large
+    double result = 0.;
+    for(i=5;i>=0;i--){
+        result += t_array[i];
+    }
+
+    double prefactor_1 = sqrt(a)*del;
+    double prefactor_2 = beta*pow(sigsq_inv*(a*del*del),-alpha);
+
+    result = prefactor_1*(1 + prefactor_2*result);
+    *zeroth_order = prefactor_1*(1+prefactor_2); //0th order term gives the barrier for efficiency
+    return result;
+}
+
+//CMF Corresponding to the Sheth Mo Tormen HMF (Sheth+ 2002)
+double dNdM_conditional_ST(double growthf, double lnM, double delta_cond, double sigma_cond){
+    double sigma1, dsigmasqdm, Barrier, factor, sigdiff_inv, result;
+    double delta_0 = delta_cond / growthf;
+    sigma1 = EvaluateSigma(lnM);
+    dsigmasqdm = EvaluatedSigmasqdm(lnM);
+    if(sigma1 < sigma_cond) return 0.;
+
+    factor = st_taylor_factor(sigma1,sigma_cond,growthf,&Barrier) - delta_0;
+
+    sigdiff_inv = sigma1 == sigma_cond ? 1e6 : 1/(sigma1*sigma1 - sigma_cond*sigma_cond);
+
+    result = -dsigmasqdm*factor*pow(sigdiff_inv,1.5)*exp(-(Barrier - delta_0)*(Barrier - delta_0)*0.5*(sigdiff_inv))/sqrt(2.*PI);
+    return result;
+}
+
+/*
+ FUNCTION dNdM_st(z, M)
+ Computes the Press_schechter mass function with Sheth-Torman correction for ellipsoidal collapse at
+ redshift z, and dark matter halo mass M (in solar masses). Moving barrier B(z,sigma) and sharp-k window functino
+
+ Uses interpolated sigma and dsigmadm to be computed faster. Necessary for mass-dependent ionising efficiencies.
+
+ The return value is the number density per unit mass of halos in the mass range M to M+dM in units of:
+ comoving Mpc^-3 Msun^-1
+
+ Reference: Sheth, Mo, Torman 2001
+ */
+double dNdlnM_st(double growthf, double lnM){
+    double sigma, dsigmadm, nuhat;
+
+    float MassBinLow;
+    int MassBin;
+
+    sigma = EvaluateSigma(lnM);
+    dsigmadm = EvaluatedSigmasqdm(lnM);
+
+    sigma = sigma * growthf;
+    dsigmadm = dsigmadm * (growthf*growthf/(2.*sigma));
+
+    nuhat = sqrt(SHETH_a) * Deltac / sigma;
+
+    return (-(cosmo_params_global->OMm)*RHOcrit) * (dsigmadm/sigma) * sqrt(2./PI)*SHETH_A * (1+ pow(nuhat, -2*SHETH_p)) * nuhat * pow(E, -nuhat*nuhat/2.0);
+}
+
+//Conditional Extended Press-Schechter Mass function, with constant barrier delta=1.682 and sharp-k window function
+double dNdM_conditional_EPS(double growthf, double lnM, double delta_cond, double sigma_cond){
+    double sigma1, dsigmasqdm, sigdiff_inv, del;
+
+    sigma1 = EvaluateSigma(lnM);
+    dsigmasqdm = EvaluatedSigmasqdm(lnM);
+
+    //limit setting
+    if(sigma1 < sigma_cond) return 0.;
+    sigdiff_inv = sigma1 == sigma_cond ? 1e6 : 1/(sigma1*sigma1 - sigma_cond*sigma_cond);
+    del = (Deltac - delta_cond)/growthf;
+
+    return -del*dsigmasqdm*pow(sigdiff_inv, 1.5)*exp(-del*del*0.5*sigdiff_inv)/sqrt(2.*PI);
+}
+
+/*
+ FUNCTION dNdM(growthf, M)
+ Computes the Press_schechter mass function at
+ redshift z (using the growth factor), and dark matter halo mass M (in solar masses).
+
+ Uses interpolated sigma and dsigmadm to be computed faster. Necessary for mass-dependent ionising efficiencies.
+
+ The return value is the number density per unit mass of halos in the mass range M to M+dM in units of:
+ comoving Mpc^-3 Msun^-1
+
+ Reference: Padmanabhan, pg. 214
+ */
+double dNdlnM_PS(double growthf, double lnM){
+    double sigma, dsigmadm;
+
+    sigma = EvaluateSigma(lnM);
+    dsigmadm = EvaluatedSigmasqdm(lnM);
+
+    sigma = sigma * growthf;
+    dsigmadm = dsigmadm * (growthf*growthf/(2.*sigma));
+    return (-(cosmo_params_global->OMm)*RHOcrit) * sqrt(2/PI) * (Deltac/(sigma*sigma)) * dsigmadm * exp(-(Deltac*Deltac)/(2*sigma*sigma));
+}
+
+//The below mass functions do not have a CMF given
+/*
+ FUNCTION dNdM_WatsonFOF(z, M)
+ Computes the Press_schechter mass function with Warren et al. 2011 correction for ellipsoidal collapse at
+ redshift z, and dark matter halo mass M (in solar masses).
+
+ The Universial FOF function (Eq. 12) of Watson et al. 2013
+
+ The return value is the number density per unit mass of halos in the mass range M to M+dM in units of:
+ comoving Mpc^-3 Msun^-1
+
+ Reference: Watson et al. 2013
+ */
+double dNdlnM_WatsonFOF(double growthf, double lnM){
+
+    double sigma, dsigmadm, f_sigma;
+
+    sigma = EvaluateSigma(lnM);
+    dsigmadm = EvaluatedSigmasqdm(lnM);
+
+    sigma = sigma * growthf;
+    dsigmadm = dsigmadm * (growthf*growthf/(2.*sigma));
+
+    f_sigma = Watson_A * ( pow( Watson_beta/sigma, Watson_alpha) + 1. ) * exp( - Watson_gamma/(sigma*sigma) );
+
+    return (-(cosmo_params_global->OMm)*RHOcrit) * (dsigmadm/sigma) * f_sigma;
+}
+
+/*
+ FUNCTION dNdM_WatsonFOF_z(z, M)
+ Computes the Press_schechter mass function with Warren et al. 2011 correction for ellipsoidal collapse at
+ redshift z, and dark matter halo mass M (in solar masses).
+
+ The Universial FOF function, with redshift evolution (Eq. 12 - 15) of Watson et al. 2013.
+
+ The return value is the number density per unit mass of halos in the mass range M to M+dM in units of:
+ comoving Mpc^-3 Msun^-1
+
+ Reference: Watson et al. 2013
+ */
+double dNdlnM_WatsonFOF_z(double z, double growthf, double lnM){
+    double sigma, dsigmadm, A_z, alpha_z, beta_z, Omega_m_z, f_sigma;
+
+    sigma = EvaluateSigma(lnM);
+    dsigmadm = EvaluatedSigmasqdm(lnM);
+
+    sigma = sigma * growthf;
+    dsigmadm = dsigmadm * (growthf*growthf/(2.*sigma));
+
+    Omega_m_z = (cosmo_params_global->OMm)*pow(1.+z,3.) / ( (cosmo_params_global->OMl) + (cosmo_params_global->OMm)*pow(1.+z,3.) + (global_params.OMr)*pow(1.+z,4.) );
+
+    A_z = Omega_m_z * ( Watson_A_z_1 * pow(1. + z, Watson_A_z_2 ) + Watson_A_z_3 );
+    alpha_z = Omega_m_z * ( Watson_alpha_z_1 * pow(1.+z, Watson_alpha_z_2 ) + Watson_alpha_z_3 );
+    beta_z = Omega_m_z * ( Watson_beta_z_1 * pow(1.+z, Watson_beta_z_2 ) + Watson_beta_z_3 );
+
+    f_sigma = A_z * ( pow(beta_z/sigma, alpha_z) + 1. ) * exp( - Watson_gamma_z/(sigma*sigma) );
+
+    return (-(cosmo_params_global->OMm)*RHOcrit) * (dsigmadm/sigma) * f_sigma;
 }
 
 double nion_fraction(double M, void *param_struct){
