+++ src/py21cmfast/src/interpolation.c
@@ -2,36 +2,39 @@
 //  We use regular grid tables since they are faster to evaluate (we always know which bin we are in)
 //  So I'm making a general function for the 1D and 2D cases
 
+#include <stdlib.h>
+#include <stdio.h>
+#include <math.h>
 #include "interpolation.h"
 #include "logger.h"
 
-void allocate_RGTable1D(int n_bin, struct RGTable1D * ptr){
+void allocate_RGTable1D(int n_bin, RGTable1D * ptr){
     ptr->n_bin = n_bin;
     ptr->y_arr = calloc(n_bin,sizeof(double));
     ptr->allocated = true;
 }
 
-void allocate_RGTable1D_f(int n_bin, struct RGTable1D_f * ptr){
+void allocate_RGTable1D_f(int n_bin, RGTable1D_f * ptr){
     ptr->n_bin = n_bin;
     ptr->y_arr = calloc(n_bin,sizeof(float));
     ptr->allocated = true;
 }
 
-void free_RGTable1D(struct RGTable1D * ptr){
+void free_RGTable1D(RGTable1D * ptr){
     if(ptr->allocated){
         free(ptr->y_arr);
         ptr->allocated = false;
     }
 }
 
-void free_RGTable1D_f(struct RGTable1D_f * ptr){
+void free_RGTable1D_f(RGTable1D_f * ptr){
     if(ptr->allocated){
         free(ptr->y_arr);
         ptr->allocated = false;
     }
 }
 
-void allocate_RGTable2D(int n_x, int n_y, struct RGTable2D * ptr){
+void allocate_RGTable2D(int n_x, int n_y, RGTable2D * ptr){
     int i;
     ptr->nx_bin = n_x;
     ptr->ny_bin = n_y;
@@ -43,7 +46,7 @@ void allocate_RGTable2D(int n_x, int n_y, struct RGTable2D * ptr){
     ptr->allocated = true;
 }
 
-void allocate_RGTable2D_f(int n_x, int n_y, struct RGTable2D_f * ptr){
+void allocate_RGTable2D_f(int n_x, int n_y, RGTable2D_f * ptr){
     int i;
     ptr->nx_bin = n_x;
     ptr->ny_bin = n_y;
@@ -55,7 +58,7 @@ void allocate_RGTable2D_f(int n_x, int n_y, struct RGTable2D_f * ptr){
     ptr->allocated = true;
 }
 
-void free_RGTable2D_f(struct RGTable2D_f * ptr){
+void free_RGTable2D_f(RGTable2D_f * ptr){
     int i;
     if(ptr->allocated){
         for(i=0;i<ptr->nx_bin;i++)
@@ -65,7 +68,7 @@ void free_RGTable2D_f(struct RGTable2D_f * ptr){
     }
 }
 
-void free_RGTable2D(struct RGTable2D * ptr){
+void free_RGTable2D(RGTable2D * ptr){
     int i;
     if(ptr->allocated){
         for(i=0;i<ptr->nx_bin;i++)
@@ -75,7 +78,7 @@ void free_RGTable2D(struct RGTable2D * ptr){
     }
 }
 
-double EvaluateRGTable1D(double x, struct RGTable1D *table){
+double EvaluateRGTable1D(double x, RGTable1D *table){
     double x_min = table->x_min;
     double x_width = table->x_width;
     int idx = (int)floor((x - x_min)/x_width);
@@ -88,7 +91,7 @@ double EvaluateRGTable1D(double x, struct RGTable1D *table){
     return result;
 }
 
-double EvaluateRGTable2D(double x, double y, struct RGTable2D *table){
+double EvaluateRGTable2D(double x, double y, RGTable2D *table){
     double x_min = table->x_min;
     double x_width = table->x_width;
     double y_min = table->y_min;
@@ -113,20 +116,17 @@ double EvaluateRGTable2D(double x, double y, struct RGTable2D *table){
 }
 
 //some tables are floats but I still need to return doubles
-double EvaluateRGTable1D_f(double x, struct RGTable1D_f *table){
+double EvaluateRGTable1D_f(double x, RGTable1D_f *table){
     double x_min = table->x_min;
     double x_width = table->x_width;
     int idx = (int)floor((x - x_min)/x_width);
     double table_val = x_min + x_width*(float)idx;
     double interp_point = (x - table_val)/x_width;
 
-    LOG_DEBUG("RG1Df: x = %.4e idx = %d val = %.4e itp %.4e",x,idx,table_val,interp_point);
-    LOG_DEBUG("Result ==> %.4e",table->y_arr[idx]*(1-interp_point) + table->y_arr[idx+1]*(interp_point));
-
     return table->y_arr[idx]*(1-interp_point) + table->y_arr[idx+1]*(interp_point);
 }
 
-double EvaluateRGTable2D_f(double x, double y, struct RGTable2D_f *table){
+double EvaluateRGTable2D_f(double x, double y, RGTable2D_f *table){
     double x_min = table->x_min;
     double x_width = table->x_width;
     double y_min = table->y_min;
